+++ src/Components/Footer/FooterData.js
@@ -26,4 +26,39 @@ const footerIconData = [
   },
 ];
 
+const footerInfoData = [
+  {
+    id: 0,
+    index_name: '법인명(상호): 주식회사 초컬리 | 사업자등록번호:261-19-23568',
+    link: '사업자정보 확인',
+  },
+  {
+    id: 1,
+    index_name:
+      '통신판매업:제 2021-서울선릉-03150호 |개인정보보호책임자:백승찬',
+    link: '',
+  },
+  {
+    id: 2,
+    index_name:
+      ' 주소:서울시 강남구 태해란로 302, 위워크 10F | 대표이사: 이예원',
+    link: '',
+  },
+  {
+    id: 3,
+    index_name: '입점문의: 채준형 | 제휴문의: ',
+    link: 'safetyHyun@choco.com',
+  },
+  {
+    id: 4,
+    index_name: '채용문의:',
+    link: 'recruit0hoon@choco.com',
+  },
+  {
+    id: 5,
+    index_name: '팩스:070-7500-0315 | 이메일:',
+    link: 'helpDongdong@choco.com',
+  },
+];
+
 export default footerIconData;
