--- src/Pages/ProductDetails/Components/GuestReview.scss
@@ -1,4 +1,0 @@
-.guest-review {
-  height: 607px;
-  background-color: rebeccapurple;
-}
