--- src/Pages/ProductDetails/Components/MoreInformation.js
@@ -1,15 +1,0 @@
-import React, { Component } from 'react';
-import './MoreInformation.scss';
-
-class MoreInformation extends Component {
-  render() {
-    return (
-      <section className="more-info-wrapper">
-        <table className="extra-info"></table>
-        <div className="why-chokurly"></div>
-      </section>
-    );
-  }
-}
-
-export default MoreInformation;
