--- src/Pages/ProductDetails/Components/MoreInformation.scss
@@ -1,13 +1,0 @@
-.more-info-wrapper {
-  width: 100%;
-  margin: 0;
-  background-color: rgb(245, 142, 24);
-  .extra-info {
-    height: 424px;
-    margin: 0;
-  }
-  .why-chokurly {
-    height: 850px;
-    padding: 44px 0 88px 0;
-  }
-}
