--- src/Pages/ProductDetails/Components/ProductImg.js
@@ -1,15 +1,0 @@
-import React, { Component } from 'react';
-import './ProductImg';
-
-class ProductImg extends Component {
-  render() {
-    return (
-      <section className="product-img">
-        <ul className="tab-groups"></ul>
-        <img src="#" alt="choKurly" />
-      </section>
-    );
-  }
-}
-
-export default ProductImg;
