--- src/Pages/ProductDetails/Components/ProductImg.scss
@@ -1,5 +1,0 @@
-.product-img {
-  height: 1000px;
-  margin: 0;
-  background-color: rgb(202, 240, 147);
-}
