--- src/Pages/ProductDetails/Components/RelatedProduct.js
@@ -1,17 +1,0 @@
-import React, { Component } from 'react';
-import './RelatedProduct.scss';
-
-class RelatedProduct extends Component {
-  render() {
-    return (
-      <section className="related-product-wrapper">
-        <div>
-          <div> </div>
-          <div className="related-product-info"></div>
-        </div>
-      </section>
-    );
-  }
-}
-
-export default RelatedProduct;
