--- src/Pages/ProductDetails/Components/RelatedProduct.scss
@@ -1,5 +1,0 @@
-.related-product-wrapper {
-  width: 1010px;
-  height: 368px;
-  background-color: powderblue;
-}
