--- src/Pages/ProductDetails/Components/ThumnailDataList.js
@@ -1,31 +1,0 @@
-const thumbnailData = [
-  {
-    id: 0,
-    info: '판매단위',
-    data: '1봉',
-    info_sub: '중량/용량',
-    data_sub: '옵션별 상이',
-  },
-  {
-    id: 2,
-    info: '배송구분',
-    data: '샛별배송/택배배송',
-  },
-  {
-    id: 3,
-    info: '포장타입',
-    data: '냉장종이포장',
-  },
-  {
-    id: 4,
-    info: '알레르기정보',
-    data: '1)초콜릿 믹스',
-    data_sub: '2) 미니넛츠 믹스:',
-  },
-  {
-    id: 5,
-    info: '상품선택',
-  },
-];
-
-export default thumbnailData;
