--- src/Pages/ProductDetails/Components/Top.scss
@@ -1,37 +1,0 @@
-.top-view {
-  width: 1050px;
-  height: 984px;
-  display: flex;
-  flex-direction: row;
-  margin: 20px auto;
-
-  .thumbnail-wrapper {
-    width: 1050px;
-    height: 552px;
-    margin: 0;
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-
-    img {
-      width: 430px;
-      height: 552px;
-    }
-
-    .thumbnail-info {
-      width: 560px;
-
-      .goods-name {
-        width: 560px;
-      }
-      .dl-list {
-        width: 560px;
-        .goods-units {
-          width: 560px;
-          display: flex;
-          flex-direction: row;
-        }
-      }
-    }
-  }
-}
