--- src/Pages/ProductDetails/ProductDetails.scss
@@ -1,1 +1,0 @@
-
